Modify the 'shortmess' option
When a search motion requires that we wrap around to the beginning or
the end of the buffer, Vim normally prints a message notifying the user
that the search is continuing from the other end of the file. This can
cause a "flicker" when working in gvim. Suppressing this message solves
the problem.