Place a :redraw call before the index is printed
By updating the screen prior to echoing the index, we avoid the
possibility of clearing the message with a delayed redrawing operation.
This seems to improve the smoothness of scanning rapidly through
searches in the GUI (e.g., by holding down 'n'). In any case, the change
doesn't hurt any functionality and can be added without worry.

See ":h :echo-redraw".